% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withloader.R
\name{marquee}
\alias{marquee}
\title{Example Shiny App}
\usage{
marquee(content, behavior = "scroll", direction = "left",
  scrollamount = 6, width = "100\%", ...)
}
\arguments{
\item{content}{The text content of the marquee.}

\item{behavior}{Sets how the text is scrolled within the marquee. Possible values are scroll, slide and alternate. If no value is specified, the default value is scroll.}

\item{direction}{Sets the direction of the scrolling within the marquee. Possible values are left, right, up and down. If no value is specified, the default value is left.}

\item{scrollamount}{Sets the amount of scrolling at each interval in pixels. The default value is 6.}

\item{width}{Sets the width in pixels or percentage value.}

\item{...}{Other parameters passed to the marquee.}

\item{height}{Sets the height in pixels or percentage value.}

\item{bgcolor}{Sets the background color through color name or hexadecimal value.}
}
\description{
Shows an example shiny app with the built-in load screens.
}
\examples{
marquee("Custom loading text here...", height=60, width=100)
}
