% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withloader.R
\name{renderCSS}
\alias{renderCSS}
\title{Example Shiny App}
\usage{
renderCSS(css, html, builtin = TRUE)
}
\arguments{
\item{css}{the name of the css file}

\item{html}{the name of the html file}

\item{builtin}{use the built-in css and html files. Default is TRUE.}
}
\description{
Shows an example shiny app with the built-in load screens.
}
