% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycustomloader-package.r
\docType{package}
\name{shinycustomloader}
\alias{shinycustomloader}
\alias{shinycustomloader-package}
\title{shinycustomloader}
\description{
This package extends the \code{shinycssloaders} R-package to allow for custom css animations
or external gifs for loading animations by wrapping a function around the shiny output.
The package contains built-in css animations such as a spinning DNA animation, walking cow animation
and a pacman animation.
}
\details{
For further reference see \url{https://github.com/emitanaka/shinycustomloader}.
}
