% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withloader.R
\name{renderCSS}
\alias{renderCSS}
\title{Render a permanent loading screen.}
\usage{
renderCSS(type, loader)
}
\arguments{
\item{type}{The type of loader to use. Possible values are image, html or text.}

\item{loader}{The name of the loader. The built in options are dnaspin, dancingtree, pacman and walkingcow.
For custom html option, the name of the css and html file have to be the same and input must be without the extension.
For custom text option, loader must be a list of marquee objects.}
}
\description{
This function follows similar to withLoader but needs no UI element. This allows for a permanent loading screen.
}
\examples{
\dontrun{marquee_list <- list(marquee("Your custom text here"))}
\dontrun{renderCSS(type="text", loader=marquee_list)}
}
